requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
PyYAML==3.12
requests==2.14.2
selenium==3.4.2
six==1.10.0
